% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crp.R
\docType{package}
\name{crp}
\alias{crp}
\alias{crp-package}
\title{Conservation Reserve Program Data}
\description{
A data package for Conservation Reserve Program data.
}
\details{
crp
A data package for Conservation Reserve Program data.
}
\seealso{
\link{crp}
}
