% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{do_format}
\alias{do_format}
\title{Format numbers as text with fixed number of decimals}
\usage{
do_format(x, digits = 2)
}
\arguments{
\item{x}{numeric vector of values to format as text}

\item{digits}{number of decimal places in the formatted output}
}
\value{
the input \code{x} formatted to text (character vector)
}
\description{
Improves the formatting of numbers to text over direct applications of
\code{round} and \code{format} by always rounding correctly and also returning the
expected number of decimals.
}
\examples{
# Simple example:
do_format(1.23456)
do_format(c(1, 2), digits = 3)

# Note that round() may loose the last decimal:
round(1.20, digits = 2)
# But do_format preserves the requested 2 decimals:
do_format(1.20, digits = 2)

# Using format() can be tricky:
format(1.206, nsmall = 2)
format(1.206, digits = 2)
# But do_format gives a useful result:
do_format(1.206, digits = 2)
}
\author{
Rune Haubo B Christensen
}
