% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_pval2}
\alias{format_pval2}
\title{Format p-values consistently}
\usage{
format_pval2(p, digits = 3)
}
\arguments{
\item{p}{Numeric p-values}

\item{digits}{Number of decimal places. P-values below \code{10^{-digits}}
are printed as e.g. \code{"< 0.001"}.}
}
\value{
Character vector
}
\description{
Easy to use modification of \code{format.pval}.
}
\examples{
format_pval2(0.03456)
format_pval2(1e-6)

}
\seealso{
\code{\link[=format.pval]{format.pval()}} which is called internally
}
\author{
Rune Haubo B Christensen
}
