% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{get_coeftab}
\alias{get_coeftab}
\title{Extract coefficient table with CI and p-value from a model fit}
\usage{
get_coeftab(
  model,
  coef_name,
  transform = function(x) x,
  ci_method = confint.default,
  digits = 3,
  digits.p = 3,
  ci_sep = " - ",
  return = c("est_ci_pval", "all")
)
}
\arguments{
\item{model}{a model fit such as that produced by \code{lm} or \code{glm}}

\item{coef_name}{name of coefficient for which to get coefficient table}

\item{transform}{optional function (e.g., \code{exp}) for a log-linear model}

\item{ci_method}{\code{method} to obtain confidence intervals}

\item{digits}{number of digits}

\item{digits.p}{number of digits in p-value}

\item{ci_sep}{CI separator}

\item{return}{return est, CI and p-value (\code{"est_ci_pval"}) or all values
(\code{"all"})}
}
\value{
a \code{data.table} with return values
}
\description{
Extract coefficient table with CI and p-value from a model fit
}
\examples{

fm <- lm(speed ~ dist, data=cars)
summary(fm)
get_coeftab(fm, "dist")

fm <- lm(Ozone ~ Solar.R + Wind + Temp, data=airquality)
summary(fm)
get_coeftab(fm, c("Solar.R", "Wind", "Temp"), ci_sep = "; ")

}
\author{
Rune Haubo B Christensen
}
