% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroag.R
\docType{package}
\name{macroag}
\alias{macroag}
\alias{macroag-package}
\title{Agricultural data for macrosystems research}
\description{
Agricultural data for macrosystems research
}
\details{
macroag
}
\seealso{
\link{macroag}
}
