% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbindall}
\alias{rbindall}
\alias{cbindall}
\title{rbind or cbind a list}
\usage{
rbindall(...)

cbindall(...)
}
\arguments{
\item{...}{objects to be \code{rbind}'ed or \code{cbind}'ed}
}
\value{
the \code{rbind}'ed or \code{cbind}'ed result
}
\description{
A generalization of \code{rbind} and \code{cbind} to allow parsing a list of objects
as argument.
}
\details{
The function definition is essentially
\code{rbindall <- function(...) do.call(rbind, ...)}
}
\examples{

x <- lapply(1:3, function(i) runif(5))
rbindall(x)
cbindall(x)

}
\author{
Rune Haubo B Christensen
}
