% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm.na}
\alias{rm.na}
\title{Remove NA from Vector}
\usage{
rm.na(x)
}
\arguments{
\item{x}{vector possibly with \code{NA} values}
}
\value{
input vector \code{x} cleaned of \code{NA} values
}
\description{
Remove NA from Vector
}
\examples{

x <- c(-1.09, NA, 1.21, -0.23, NA, 0.31, -0.28)
rm.na(x)

}
\author{
Rune Haubo B Christensen
}
