% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table_utils.R
\name{show_na}
\alias{show_na}
\title{Summarize \code{NA} values in a \code{data.table}}
\usage{
show_na(d)
}
\arguments{
\item{d}{a \code{data.table} or \code{data.frame}}
}
\value{
A \code{matrix} summarizing the number of \code{NA}s and zero-length characters
}
\description{
Summarize \code{NA} values in a \code{data.table}
}
\examples{

show_na(airquality)
show_na(iris)

}
\author{
Rune Haubo B Christensen
}
