% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_set}
\alias{show_set}
\title{Detail union and intersection of two variables}
\usage{
show_set(A, B, unique = TRUE)
}
\arguments{
\item{A}{a vector}

\item{B}{a vector}

\item{unique}{compare only unique entries in \code{A} and \code{B}?}
}
\value{
a \code{data.frame} with 1 row and columns \code{nA}, \code{nB}, \code{uniqueA}, \code{uniqueB}, \code{union} \code{intersect}, \code{AnotB}, \code{BnotA}
}
\description{
Convenience function to compare two variables in terms of length, number of unique values, union, and intersection.
}
\examples{

head(iris)
with(iris, show_set(Sepal.Length, Petal.Length))

}
\author{
Rune Haubo B Christensen
}
