% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroag.R
\docType{data}
\name{tillage_ctic}
\alias{tillage_ctic}
\title{Conservation tillage data}
\description{
Conservation tillage data
}
\details{
tillage_ctic
}
\references{
\url{https://water.usgs.gov/GIS/metadata/usgswrd/XML/ds573_tillage_lu01.xml}

Baker, N.T. 2011. Tillage practices in the conterminous United States, 1989–2004: Datasets aggregated by watershed. U.S. Geological Survey Data Series 573. USGS, Reston, VA
}
\keyword{datasets}
