% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table_utils.R
\name{totpct}
\alias{totpct}
\title{Add column total and row percentages to a \code{data.table} tabulation}
\usage{
totpct(tab, name_col = 1, n_col = 2, digits = 2)
}
\arguments{
\item{tab}{a \code{data.table} tabulation; see Examples}

\item{name_col}{name or index of the label column (character or integer)}

\item{n_col}{name or index of the count/numeric column (character or integer)}

\item{digits}{number of decimal places in the percentage column}
}
\value{
\code{tab} with appended total-row and percent-column
}
\description{
Add column total and row percentages to a \code{data.table} tabulation
}
\examples{
Iris <- as.data.table(iris)

tab <- Iris[, .N, Species]
totpct(tab)

tab <- Iris[, .(Mean = mean(Sepal.Length), .N), Species]
totpct(tab, n_col = 3, digits=1)
}
\author{
Rune Haubo B Christensen
}
