% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crp.R
\docType{package}
\name{crp}
\alias{crp}
\alias{crp-package}
\title{Conservation Reserve Program data}
\description{
Conservation Reserve Program dataset
}
\details{
crp
Conservation Reserve Program dataset
}
\seealso{
\link{crp}
}
