% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroag.R
\docType{data}
\name{crp}
\alias{crp}
\title{Conservation Reserve Program}
\description{
Data from the Conservation Reserve Program
}
\details{
crp
}
\references{
\url{https://www.fsa.usda.gov/programs-and-services/conservation-programs/reports-and-statistics/conservation-reserve-program-statistics/index}
}
\keyword{datasets}
