% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroag.R
\docType{data}
\name{crp}
\alias{crp}
\title{Conservation Reserve Program}
\description{
Data from the Conservation Reserve Program
}
\details{
crp
}
\references{
\url{https://www.fsa.usda.gov/programs-and-services/conservation-programs/reports-and-statistics/conservation-reserve-program-statistics/index}

Sprague, L. A., and J. A. M. Gronberg. 2012. Relating Management Practices and Nutrient Export in Agricultural Watersheds of the United States. Journal of Environment Quality 41: 1939. doi:10.2134/jeq2012.0073
}
\keyword{datasets}
