% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{do_format}
\alias{do_format}
\title{Format numbers as text with fixed number of decimals}
\usage{
do_format(x, digits = 2)
}
\arguments{
\item{x}{numeric vector of values to format as text}

\item{digits}{number of digits in the formatted number}
}
\value{
the input \code{x} formatted to text (character vector)
}
\description{
Improves the formatting of numbers to text over direct applications of
\code{round} and \code{format} as illustrated in examples.
}
\examples{
# Simple example:
do_format(1.23456)
do_format(c(1, 2), 3)

# Note that round() looses the last digit:
do_format(1.20)
round(1.20, digits=2)

# Note that here format() doesn't round correctly:
do_format(1.206)
format(1.206, digits=2)
}
