% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{first}
\alias{first}
\alias{last}
\title{Get First or Last Element}
\usage{
first(x)

last(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the first or last element
}
\description{
Get First or Last Element
}
\examples{

x <- rnorm(5)
x
first(x)
last(x)

}
\author{
Rune Haubo B Christensen
}
