% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_estci}
\alias{format_estci}
\title{Format estimate with confidence interval}
\usage{
format_estci(est, lwr, upr, digits = 2, ci_sep = " - ")
}
\arguments{
\item{est}{parameter estimate (numeric scalar)}

\item{lwr}{lower limit (numeric scalar)}

\item{upr}{upper limit (numeric scalar)}

\item{digits}{number of digits (numeric scalar)}

\item{ci_sep}{CI separator}
}
\value{
The text "estimate (lower - upper)" (character vector)
}
\description{
Format as "estimate (lower - upper)"
}
\examples{

# Basic usage:
format_estci(1.2, 0.8, 1.6)

# Change CI-separator:
format_estci(1.2, 0.8, 1.6, ci_sep="; ")

# Multiple point and interval estimates:
format_estci(1:2, 0:1, 2:3)

}
\seealso{
\code{\link[=do_format]{do_format()}} used for the formatting of estimate, lower and upper
}
