% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_estci}
\alias{format_estci}
\title{Format estimate with confidence interval}
\usage{
format_estci(
  est,
  lwr,
  upr,
  digits = 2,
  ci_sep = " - ",
  use_exp = FALSE,
  upr_lim = 100
)
}
\arguments{
\item{est}{parameter estimate (numeric scalar)}

\item{lwr}{lower limit (numeric scalar)}

\item{upr}{upper limit (numeric scalar)}

\item{digits}{number of digits (numeric scalar)}

\item{ci_sep}{CI separator}

\item{use_exp}{exp-transform \code{est}, \code{lwr}, and \code{upr}}

\item{upr_lim}{upper limit for printing of \code{upr} applied when \code{use_exp} is \code{TRUE} to avoid very large numbers.}
}
\value{
The text "estimate (lower - upper)" (character vector)
}
\description{
Format as "estimate (lower - upper)"
}
\examples{

# Basic usage:
format_estci(1.2, 0.8, 1.6)

# Change CI-separator:
format_estci(1.2, 0.8, 1.6, ci_sep="; ")

# Multiple point and interval estimates:
format_estci(1:2, 0:1, 2:3)

# Illustrate use of use_exp argument:
format_estci(1:2, 0:1, 2:3, use_exp = TRUE)

# When use_exp = TRUE the upper limit is capped at upr_lim = 100: 
format_estci(1:2, 0:1, 4:5, use_exp = TRUE)

}
\seealso{
\code{\link[=do_format]{do_format()}} used for the formatting of estimate, lower and upper
}
\author{
Rune Haubo B Christensen
}
