% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table_utils.R
\name{totpct}
\alias{totpct}
\title{Add Column Total and Row Percentages to a \code{data.table} Tabulation}
\usage{
totpct(tab, name_col = 1, n_col = 2, digits = 2)
}
\arguments{
\item{tab}{a \code{data.table} tabulation; see Examples}

\item{name_col}{indicator of name column (character or numeric)}

\item{n_col}{indicator of the numeric column (character or numeric)}

\item{digits}{number of digits in percentages in result}
}
\value{
\code{tab} with appended total-row and percent-column
}
\description{
Add Column Total and Row Percentages to a \code{data.table} Tabulation
}
\examples{

Iris <- as.data.table(iris)

tab <- Iris[, .N, Species]
totpct(tab)

tab <- Iris[, .(Mean = mean(Sepal.Length), .N), Species]
totpct(tab, n_col = 3, digits=1)

}
\author{
Rune Haubo B Christensen
}
